requirements.txt
black==24.4.2
Django==4.2.11
django-environ==0.10.0
python-upwork-oauth2==3.1.0
psycopg2-binary==2.8.5
sentry-sdk==1.17.0